Correlation chart improvements
