+++ src/components/index.ts
@@ -5,3 +5,4 @@ export * from "./tags";
 export * from "./map-info";
 export * from "./lithology";
 export * from "./buttons";
+export { buildCrossSectionLayers } from "~/_utils/map-layers";
