+++ src/pages/correlation/main.module.sass
@@ -1,11 +1,13 @@
+$map-size: 40em
+
 .column-selection-map
-  height: 30em
+  height: $map-size
   position: relative
   :global(.main-view)
     height: 100%
 
 .assistant
-  width: 30em
+  width: $map-size
 
 .correlation-diagram
   flex-grow: 1
