+++ src/pages/map/map-interface/map-page/map-styles/overlay.ts
@@ -1,3 +1,5 @@
+import { buildCrossSectionLayers } from "~/_utils/map-layers";
+
 /** Add extra types we use in this style... */
 interface SourceExt extends mapboxgl.Source {
   cluster?: boolean;
@@ -78,67 +80,6 @@ const overlaySources: { [k: string]: SourceExt } = {
     },
   },
 };
-
-export function buildCrossSectionLayers(): mapboxgl.Layer[] {
-  // Get CSS colors from settings
-  const ruleColor = getComputedStyle(document.body).getPropertyValue(
-    "--panel-background-color"
-  );
-
-  const centerColor = getComputedStyle(document.body).getPropertyValue(
-    "--panel-rule-color"
-  );
-
-  const crossSectionPointPaint = {
-    "circle-radius": {
-      stops: [
-        [0, 3],
-        [12, 5],
-      ],
-    },
-    "circle-color": centerColor,
-    "circle-stroke-width": {
-      stops: [
-        [0, 2],
-        [12, 4],
-      ],
-    },
-    "circle-stroke-color": ruleColor,
-  };
-
-  return [
-    {
-      id: "crossSectionLine",
-      type: "line",
-      source: "crossSectionLine",
-      paint: {
-        "line-width": {
-          stops: [
-            [0, 1],
-            [12, 3],
-          ],
-        },
-        "line-color": ruleColor,
-        "line-opacity": 1,
-      },
-    },
-    {
-      id: "crossSectionEndpoint",
-      type: "circle",
-      source: "crossSectionEndpoints",
-      paint: crossSectionPointPaint,
-    },
-    {
-      id: "elevationMarker",
-      type: "circle",
-      source: "elevationMarker",
-      paint: {
-        ...crossSectionPointPaint,
-        "circle-color": "#4bc0c0",
-      },
-    },
-  ];
-}
 
 export function buildOverlayLayers(): mapboxgl.Layer[] {
   // Get CSS colors from settings
